app/backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cryptography==2.9.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==0.16.1
wincertstore==0.2
zipp==3.1.0
